requirements.txt
boto3
snowflake-connector
pyspark
